refactor: 매장 상품 검색 페이지 수정
